Move the Interp typeclass into a different module
