+++ gibbon-compiler/src/Gibbon/GenericOps.hs
@@ -6,10 +6,9 @@
 -- representations inside the compiler.
 
 module Gibbon.GenericOps
-    ( Interp(..)
-
+    (
     -- * Free Variables, generic interface
-    , FreeVars(..)
+      FreeVars(..)
 
     -- * Genereric interface for expressions and select passes
     , Expression(..), NoExt
@@ -23,35 +22,9 @@ import Control.DeepSeq (NFData)
 import Data.Loc
 import GHC.Generics
 import Text.PrettyPrint.GenericPretty
-import System.IO.Unsafe (unsafePerformIO)
 import qualified Data.Set as S
-import qualified Data.ByteString.Lazy.Char8 as B
 
 import Gibbon.Common
-import Gibbon.Interp (Value)
-
---------------------------------------------------------------------------------
--- Things which can be interpreted to yield a final, printed value.
---------------------------------------------------------------------------------
-
--- | Pure Gibbon programs, at any stage of compilation, should always
--- be evaluatable to a unique value.  The only side effects are timing.
-class Interp a where
-  {-# MINIMAL interpProg #-}
-
-  interpProg :: RunConfig -> a -> IO (Value, B.ByteString)
-
-  -- | Interpret while ignoring timing constructs, and dropping the
-  -- corresponding output to stdout.
-  interpNoLogs :: RunConfig -> a -> String
-  interpNoLogs rc p = unsafePerformIO $ show . fst <$> interpProg rc p
-
-  -- | Interpret and produce a "log" of output lines, as well as a
-  -- final, printed result.  The output lines include timing information.
-  interpWithStdout :: RunConfig -> a -> IO (String,[String])
-  interpWithStdout rc p = do
-   (v,logs) <- interpProg rc p
-   return (show v, lines (B.unpack logs))
 
 -------------------------------------------------------------------------------
 -- Free Variables
